Create a job to fetch deals from pipeline
Here I've set up sidekiq to handle asyncronous task running, and added
it to my Procfile.

A new ActiveJob has been created, `FetchDealsJob`, to query the
pipeline api. In this job, each returned deal instance and associated
stage is extracted into their own arrays, and bulk upserted in
sequence for efficiency.

A test has been added to confirm that records are being created as
expected, with the API call itself stubbed via webmock to keep the
tests in isolation.

Additionally, `sidekiq-scheduler` has been added and configured to run
this job every minute. Note that this is probably much more often than
I would set it to run in production, but as a proof of concept it
gives more iterations to evaluate.