Implement deal stage api and frontend
This commit adds an api namespace to the app, with a single endpoint
to return deal stages. Simple json rendering is used in lieu of a more
complicated serializer solution to keep the scope of this exam small.

`DealStage` has been updated with a method to sum the value of all
associated deals, and `DealStage#as_json` has been added to format the
instances properly.

Tests were added to ensure that the api endpoint functions, is
properly ordered, and contains a correct total value.

On the frontend, a simple chart was constructed using
recharts (http://recharts.org/en-US). A single React component was
used, and internal state was chosen in lieu of a more complicated
redux store; again, to keep the scope of this project from bloating.

Data is loaded in on component mount via the `fetch` api.
Traditionally I would opt for something more robust and browser
compatible, such as Axios, but it did not seem necessary for the scope
of this problem.

Simple styling was added to keep the page from looking overly bland.